refactor app for better organization and prod ready structure
