+++ bps_internal_tools/admin/__init__.py
@@ -1,7 +1,6 @@
 from flask import Blueprint
-admin_bp = Blueprint("admin", __name__, template_folder="../templates")
+admin_bp = Blueprint("admin", __name__)
 TOOL_SLUG = "admin"
-
 
 # IMPORTANT: import routes so the @toc_bp.route decorators execute
 from . import routes  # noqa: E402, F401
