+++ bps_internal_tools/extensions.py
@@ -1,0 +1,8 @@
+# bps_internal_tools/extensions.py
+from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy.orm import DeclarativeBase
+
+class Base(DeclarativeBase):
+    pass
+
+db = SQLAlchemy(model_class=Base)
