+++ bps_internal_tools/models.py
@@ -1,4 +1,4 @@
-# models.py
+from .extensions import db
 from sqlalchemy import (
     Column, String, Integer, Boolean, DateTime, ForeignKey, UniqueConstraint, Text
 )
@@ -7,7 +7,7 @@ from sqlalchemy.orm import declarative_base, relationship
 Base = declarative_base()
 
 # --- Auth / RBAC ---
-class Role(Base):
+class Role(db.Model):
     __tablename__ = "roles"
     id = Column(Integer, primary_key=True)
     name = Column(String(64), unique=True, nullable=False)     # e.g. 'admin', 'attendance'
@@ -15,7 +15,7 @@ class Role(Base):
 
     tools = relationship("RoleTool", back_populates="role", cascade="all, delete-orphan")
 
-class RoleTool(Base):
+class RoleTool(db.Model):
     __tablename__ = "role_tools"
     id = Column(Integer, primary_key=True)
     role_id = Column(Integer, ForeignKey("roles.id", ondelete="CASCADE"), nullable=False)
@@ -24,7 +24,7 @@ class RoleTool(Base):
 
     role = relationship("Role", back_populates="tools")
 
-class User(Base):
+class User(db.Model):
     __tablename__ = "users_auth"
     id = Column(Integer, primary_key=True, autoincrement=True)
     username = Column(String(128), unique=True, nullable=False)
@@ -35,7 +35,7 @@ class User(Base):
 
     
 # --- Canvas-like data ---
-class Course(Base):
+class Course(db.Model):
     __tablename__ = "courses"
     course_id = Column(String(32), primary_key=True)      # 'c003936' etc.
     integration_id = Column(String(64))
@@ -49,7 +49,7 @@ class Course(Base):
     course_format = Column(String(64))
     blueprint_course_id = Column(String(64))
 
-class People(Base):
+class People(db.Model):
     __tablename__ = "users_canvas"
     user_id = Column(String(64), primary_key=True)        # Canvas user_id
     integration_id = Column(String(64))
@@ -65,7 +65,7 @@ class People(Base):
     status = Column(String(64))
     pronouns = Column(String(255))
 
-class Enrollment(Base):
+class Enrollment(db.Model):
     __tablename__ = "enrollments"
     id = Column(Integer, primary_key=True, autoincrement=True)
     course_id = Column(String(32), ForeignKey("courses.course_id", ondelete="CASCADE"), index=True)
