+++ bps_internal_tools/static/app.js
