+++ bps_internal_tools/static/styles.css
@@ -1,6 +1,4 @@
 /* ===== Base ===== */
-@import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;500;600;700;800&display=swap');
-
 :root {
   --bg: #0b1020;
   --panel: #111833;
