+++ bps_internal_tools/static/ui.js
