+++ bps_internal_tools/templates/select_course.html
