+++ bps_internal_tools/templates/take_attendance.html
