+++ requirements.txt
@@ -11,3 +11,5 @@ gspread_formatting
 SQLAlchemy>=2.0
 alembic>=1.13
 pymysql>=1.1 
+flask_sqlalchemy
+gunicorn
