+++ wsgi.py
@@ -1,0 +1,2 @@
+from bps_internal_tools import create_app
+app = create_app()
